feat(core): implement stack-agnostic tree-sitter parser factory
