+++ src/aio_agentic_sdlc/parsers/base.py
@@ -1,0 +1,5 @@
+from typing import Any
+
+class BaseFileParser:
+    def parse(self, generator: Any, file_path: str):
+        raise NotImplementedError
